correzzioni generiche

+++ ft_strlen.c
@@ -12,7 +12,7 @@
 
 #include <stdio.h>
 
-int	ft_strlen(char *s)
+size_t	ft_strlen(char *s)
 {
 	int	i;
